Move View class to the browser module
