+++ src/collective/polls/browser/__init__.py
@@ -1,7 +1,14 @@
 # -*- coding: utf-8 -*-
 from AccessControl import Unauthorized
+from Acquisition import aq_inner
+from Acquisition import aq_parent
+from collective.polls import MessageFactory as _
 from collective.polls.polls import IPolls
 from plone import api
+from Products.CMFCore.interfaces import ISiteRoot
+from Products.Five.browser import BrowserView
+from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
+from zope.component import getMultiAdapter
 from zope.component import queryUtility
 
 
@@ -83,3 +90,88 @@ class PollsViewMixin:
         """Return the number of votes so far."""
         poll = self.poll()
         return poll.total_votes
+
+
+class View(PollsViewMixin, BrowserView):
+
+    index = ViewPageTemplateFile('templates/view.pt')
+
+    def render(self):
+        return self.index()
+
+    def __call__(self):
+        self.update()
+        return self.render()
+
+    def update(self):
+        context = aq_inner(self.context)
+        self.context = context
+        self.state = getMultiAdapter(
+            (context, self.request), name=u'plone_context_state')
+        self.wf_state = self.state.workflow_state()
+        # Handle vote
+        form = self.request.form
+        self.errors = []
+        self.messages = []
+
+        # if the poll is open and anonymous should vote but the parent folder
+        # is private.. inform the user.
+
+        # When the poll's container is the site's root, we do not need to
+        # check the permissions.
+        container = aq_parent(aq_inner(self.context))
+
+        if 'open' == self.wf_state and not ISiteRoot.providedBy(container):
+            roles = [
+                r['name'] for r in
+                self.context.rolesOfPermission('collective.polls: Vote')
+                if r['selected']]
+
+            if 'Anonymous' not in roles and self.context.allow_anonymous:
+                msg = _(
+                    u"Anonymous user won't be able to vote, you forgot to "
+                    u'publish the parent folder, you must sent back the poll '
+                    u'to private state, publish the parent folder and open '
+                    u'the poll again')
+                api.portal.show_message(msg, self.request, type='warn')
+
+        if 'poll.submit' in form:
+            self._updateForm(form)
+        # Update status messages
+        for error in self.errors:
+            api.portal.show_message(error, self.request, type='warn')
+        for msg in self.messages:
+            api.portal.show_message(msg, self.request, type='info')
+
+        # XXX
+        # if 'voting.from' in form:
+        #     url = form['voting.from']
+        #     self.request.RESPONSE.redirect(url)
+
+    def _updateForm(self, form):
+        INVALID_OPTION = _(u'Invalid option')
+        options = form.get('options', '')
+        if isinstance(options, list):
+            self.errors.append(INVALID_OPTION)
+        elif isinstance(options, str):
+            if not options.isdigit():
+                self.errors.append(INVALID_OPTION)
+            else:
+                options = int(options)
+        if not self.errors:
+            # Let's vote
+            try:
+                self.context.setVote(options, self.request)
+                self.messages.append(_(u'Thanks for your vote'))
+                # We do this to avoid redirecting anonymous user as
+                # we just sent them the cookie
+                self._has_voted = True
+            except Unauthorized:
+                self.errors.append(_(u'You are not authorized to vote'))
+
+    def poll(self):
+        return self.context
+
+    def getOptions(self):
+        """Return available options."""
+        return self.context.getOptions()
