+++ src/collective/polls/browser/configure.zcml
@@ -5,7 +5,7 @@
   <browser:page
       name="view"
       for="collective.polls.content.poll.IPoll"
-      class="..content.poll.View"
+      class=".View"
       permission="zope2.View"
       />
