+++ src/collective/polls/browser/templates/view.pt
